+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - kube-proxy: explicitly set cluster-cidr
 - kube-proxy: simplify proxy-mode configuration
 - Make kubernetes services network route ipvs compatible
+- Packages: install `ipvsadm` and `ipset` for IPVS proxy mode
 
 # 1.11.3 (11.07.2018)
 - CNI plugins: bump to 0.7.1